+++ include/render/GpuPrimaryPathDescriptor.h
@@ -14,6 +14,7 @@ namespace render {
   inline constexpr std::uint32_t gpuPrimaryPathGenerationModeEquirectangular = 4u;
   inline constexpr std::uint32_t gpuPrimaryPathGenerationModeSpherical = 5u;
   inline constexpr std::uint32_t gpuPrimaryPathGenerationModeFishEye = 6u;
+  inline constexpr std::uint32_t gpuPrimaryPathGenerationModeTiltShift = 7u;
 
   struct alignas(16) GpuRectilinearPrimaryPathDescriptor {
     std::array<float, 4> originOrDirection{};